refactor: import Missing* exceptions from pint.exceptions
Stop re-exporting them from timing_model; point remaining call sites
at the canonical module.
